+++ internal/checks/promql_syntax.go
@@ -39,7 +39,7 @@ func (c SyntaxCheck) Reporter() string {
 	return SyntaxCheckName
 }
 
-func (c SyntaxCheck) Check(_ context.Context, _ string, rule parser.Rule, _ []discovery.Entry) (problems []Problem) {
+func (c SyntaxCheck) Check(_ context.Context, _ discovery.Path, rule parser.Rule, _ []discovery.Entry) (problems []Problem) {
 	expr := rule.Expr()
 	if expr.SyntaxError != nil {
 		problems = append(problems, Problem{